amax, max: cleaned up mess and proposing a dual generic API for functions that can return a scalar or an array based on axis param
