+++ src/NumSharp.Core/APIs/np.manipulation.cs
@@ -8,5 +8,7 @@ namespace NumSharp
     {
         public static int roll(NDArray nd, int shift, int axis = -1)
             => axis == -1 ? nd.roll(shift) : nd.roll(shift, axis);
+
+        public static NDArray ravel(NDArray a) => a.ravel();
     }
 }
