--- src/NumSharp.Core/APIs/np.searching.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace NumSharp
-{
-    public static partial class np
-    {
-        public static int argmax(NDArray nd)
-            => nd.argmax();
-    }
-}
