+++ src/NumSharp.Core/Math/NdArray.AMin.cs
