--- src/NumSharp.Core/Math/np.amax.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace NumSharp
-{
-    public static partial class np
-    {
-        public static NDArray amax(NDArray nd, int? axis = null)
-        {
-            return nd.amax(axis);
-        }
-    }
-}
