+++ test/NumSharp.UnitTest/LinearAlgebra/NdArray.QR.Test.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using NumSharp.Extensions;
@@ -26,7 +26,7 @@ namespace NumSharp.UnitTest.LinearAlgebra
             var nd2 = np.dot(Q.transpose(), R);
 
             // make sure the highest difference is lower than 0.0000000001
-            Assert.IsTrue( ((double)(nd1 - nd2).absolute().max()[0] )  < 0.00000000001); 
+            Assert.IsTrue( ((double)(nd1 - nd2).absolute().max<double>() )  < 0.00000000001); 
         }
         
     }
